oncall/my_metrics/requirements.txt
requests~=2.31.0
prometheus-client~=0.17.1
PyYAML~=6.0.1